requirements.txt
streamlit==1.33.0
pdf2image==1.17.0
ultralytics==8.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0